pyproject.toml
[project]
name = "ezflow-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.5",
    "fastapi[standard]>=0.128.0",
    "langchain-community>=0.4.1",
    "langchain-mongodb>=0.10.0",
    "langchain-openai>=1.1.7",
    "pydantic>=2.12.5",
    "pyodbc>=5.3.0",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.45",
]
